requirements.txt
asgiref==3.8.1
click==8.1.8
colorama==0.4.6
dj-database-url==2.3.0
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
gunicorn==23.0.0
h11==0.14.0
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.34.0
whitenoise==6.8.2
